coding-agent/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "An intelligent coding assistant powered by LangChain and Ollama"
readme = "docs/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "coding-assistant", "langchain", "ollama", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.1.20",
    "langchain-community>=0.0.38",
    "langchain-core>=0.1.52",
    "ollama>=0.1.9",
    "chromadb>=0.4.24",
    "sentence-transformers>=2.6.1",
    "sqlalchemy>=2.0.29",
    "mysql-connector-python>=8.3.0",
    "pymysql>=1.1.0",
    "tree-sitter>=0.21.3",
    "tree-sitter-languages>=1.10.2",
    "python-dotenv>=1.0.1",
    "gitignore-parser>=0.1.11",
    "watchdog>=4.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "rich>=13.7.1",
    "click>=8.1.7",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "black>=24.3.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
]

[project.scripts]
coding-agent = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"
